chore(azure): Outputs options in the provider
