+++ prowler/providers/common/outputs.py
@@ -74,33 +74,6 @@ class Provider_Output_Options:
                     makedirs(arguments.output_directory + "/compliance", exist_ok=True)
 
 
-class Azure_Output_Options(Provider_Output_Options):
-    def __init__(self, arguments, identity, mutelist_file, bulk_checks_metadata):
-        # First call Provider_Output_Options init
-        super().__init__(arguments, mutelist_file, bulk_checks_metadata)
-
-        # Confire Shodan API
-        # TODO: review shodan for the new AWS provider
-        # if arguments.shodan:
-        #     audit_info = change_config_var(
-        #         "shodan_api_key", arguments.shodan, audit_info
-        #     )
-
-        # Check if custom output filename was input, if not, set the default
-        if (
-            not hasattr(arguments, "output_filename")
-            or arguments.output_filename is None
-        ):
-            if identity.domain != "Unknown tenant domain (missing AAD permissions)":
-                self.output_filename = (
-                    f"prowler-output-{identity.domain}-{output_file_timestamp}"
-                )
-            else:
-                self.output_filename = f"prowler-output-{'-'.join(identity.tenant_ids)}-{output_file_timestamp}"
-        else:
-            self.output_filename = arguments.output_filename
-
-
 class Gcp_Output_Options(Provider_Output_Options):
     def __init__(self, arguments, identity, mutelist_file, bulk_checks_metadata):
         # First call Provider_Output_Options init
